chore: move default_settings to constants
