+++ src/common/constants.ts
@@ -2,6 +2,8 @@
 // Could be split up into i.e.
 // constants/messages
 // constants/[...etc]
+import i18n from "i18next";
+import type { SettingsStorage } from "~/types";
 
 export const ABORT_PROMPT_ERROR = "Prompt was closed";
 export const USER_REJECTED_ERROR = "User rejected";
@@ -168,3 +170,20 @@ export enum TIPS {
   DEMO = "demo",
   ADDRESS = "address",
 }
+
+export const DEFAULT_SETTINGS: SettingsStorage = {
+  browserNotifications: true,
+  websiteEnhancements: true,
+  legacyLnurlAuth: false,
+  isUsingLegacyLnurlAuthKey: false,
+  userName: "",
+  userEmail: "",
+  locale: i18n.resolvedLanguage,
+  theme: "system",
+  showFiat: true,
+  currency: CURRENCIES.USD,
+  exchange: "alby",
+  debug: false,
+  nostrEnabled: false,
+  closedTips: [],
+};
