+++ tests/fixtures/settings.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_SETTINGS } from "~/extension/background-script/state";
+import { DEFAULT_SETTINGS } from "~/common/constants";
 import type { SettingsStorage } from "~/types";
 
 export const settingsFixture: SettingsStorage = { ...DEFAULT_SETTINGS };
